refactor: clean-up & folder
