+++ shared/routes/index.ts
@@ -2,9 +2,7 @@ import { ConditionalExcept, EmptyObject, Jsonify } from "type-fest"
 import z, { ZodType } from "zod"
 
 import { zApplicationRoutes } from "./application.routes"
-import { zApplicationRoutesV2 } from "./application.routes.v2"
 import { zAppointmentsRoute } from "./appointments.routes"
-import { zAppointmentsRouteV2 } from "./appointments.routes.v2"
 import { IRouteSchema, IRouteSchemaWrite } from "./common.routes"
 import { zCoreRoutes } from "./core.routes"
 import { zEligibleTrainingsForAppointmentRoutes } from "./eligibleTrainingsForAppointment.routes"
@@ -12,7 +10,6 @@ import { zEmailsRoutes } from "./emails.routes"
 import { zEtablissementRoutes } from "./etablissement.routes"
 import { zFormationRoute } from "./formations.routes"
 import { zFormulaireRoute } from "./formulaire.route"
-import { zJobsRoutesV2 } from "./jobs.routes.v2"
 import { zLoginRoutes } from "./login.routes"
 import { zMetiersRoutes } from "./metiers.routes"
 import { zMetiersDAvenirRoutes } from "./metiersdavenir.routes"
@@ -30,6 +27,9 @@ import { zV1FormationsRoutes } from "./v1Formations.routes"
 import { zV1FormationsParRegion } from "./v1FormationsParRegion.routes"
 import { zV1JobsRoutes } from "./v1Jobs.routes"
 import { zV1JobsEtFormationsRoutes } from "./v1JobsEtFormations.routes"
+import { zApplicationRoutesV2 } from "./v2/application.routes.v2"
+import { zAppointmentsRouteV2 } from "./v2/appointments.routes.v2"
+import { zJobsRoutesV2 } from "./v2/jobs.routes.v2"
 import { zJobsRoutesV3 } from "./v3/jobs/jobs.routes.v3"
 
 const zRoutesGetP1 = {
