+++ shared/routes/v2/appointments.routes.v2.ts
@@ -1,8 +1,7 @@
-import { referrers } from "../constants/referers"
-import { extensions } from "../helpers/zodHelpers/zodPrimitives"
-import { z } from "../helpers/zodWithOpenApi"
-
-import { IRoutesDef } from "./common.routes"
+import { referrers } from "../../constants/referers"
+import { extensions } from "../../helpers/zodHelpers/zodPrimitives"
+import { z } from "../../helpers/zodWithOpenApi"
+import { IRoutesDef } from "../common.routes"
 
 const ZAppointmentContextParcoursup = z
   .object({
